refactor: split webdav.py into webdav_discovery + webdav_match (iter28)
Decompose the webdav.py god-file: move PROPFIND scan/recurse internals into
webdav_discovery.py (415 NLOC) and episode/title tagging+scoring into webdav_match.py
(101 NLOC, a true leaf). webdav.py NLOC 720->241; all three modules now <500.

Net-positive split (learned from a reverted first attempt that left a 536-NLOC module
and added params): both new modules are <500, and the recursion's call back into
find_video_file uses a function-local `import resources.lib.webdav as _webdav`
(NOT a threaded param) -> no new params>8, no top-level import cycle, and @patch
resolves at call time. urlparse kept function-local to preserve a call-count test.
All test-@patched names stayed in webdav.py; moved test/filter-accessed names
(_episode_tags etc.) re-exported via __all__.

Codacy complexity 28 -> 27 (file-nloc>500: 5->4; params>8 unchanged at 16). Verified:
2138/2140 tests pass (the 2 are pre-existing wall-clock timing flakes that pass with
free CPU and fail at baseline regardless), pylint 10/10, ruff+black clean, vermin 3.8.
